+++ packages/web/src/contexts/providers/SettingsProvider.tsx
@@ -1,5 +1,11 @@
 import React from "react";
-import { DEFAULT_CONTEXT, DEFAULT_DRY, DEFAULT_MODEL } from "gpt-turbo";
+import {
+    DEFAULT_CONTEXT,
+    DEFAULT_DISABLEMODERATION,
+    DEFAULT_DRY,
+    DEFAULT_MODEL,
+    DEFAULT_STREAM,
+} from "gpt-turbo";
 
 import { SettingsContext, SettingsContextValue } from "../SettingsContext";
 import useStorage from "../../hooks/useStorage";
@@ -21,8 +27,8 @@ export default ({ children }: SettingsProviderProps) => {
             model: DEFAULT_MODEL,
             context: DEFAULT_CONTEXT,
             dry: DEFAULT_DRY,
-            disableModeration: "on",
-            stream: true,
+            disableModeration: DEFAULT_DISABLEMODERATION,
+            stream: DEFAULT_STREAM,
             save: false,
         },
         settingsSchema
